requirements.txt
async-timeout==3.0.1
discord.py==1.6.0
aPRAW==0.6.8a0
prawcore==2.0.0
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.17.0
pyttsx3==2.90
youtube-dl==2021.2.4.1
youtube-dlc==2020.11.11.post3
